1. Copy the az cli's cache directory from the host to the devcontainer. This means that you no longer have to do `az login` in a new devcontainer if you are logged in on the host.
1. Allow a simple `make` command to deploy and run all tests on a new environment.
1. Separate out go E2E tests from eminence, so you can go `make e2e` and not run all the eminence tests.
1. `ensure-environment.sh` now stores a hash of the environment definition as a tag on the resource group, allowing us to skip deployment if the environment has not changed.

Here is a `make` command on a brand-new devcontainer with no existing environment (75x video speedup):

![full-1 (1).gif](https://dev.azure.com/msresearch/b54927b4-7877-4e5f-b8fc-226850558453/_apis/git/repositories/9652e2d6-5c0f-4ada-ae48-53e308ca25c2/pullRequests/64410/attachments/full-1%20%281%29.gif)

+++ Makefile
@@ -2,44 +2,47 @@
 SHELL = /bin/bash
 .SHELLFLAGS = -ecuo pipefail
 
-SERVER_PATH=server/Tyger.Server
+.DEFAULT_GOAL := test
 
+# trick to lazily evaluate this at most once: https://make.mad-scientist.net/deferred-simple-variable-expansion/
+ENVIRONMENT_CONFIG = $(eval ENVIRONMENT_CONFIG := $$(shell scripts/get-context-environment-config.sh))$(ENVIRONMENT_CONFIG)
+
+SERVER_PATH=server/Tyger.Server
 SECURITY_ENABLED=true
 DEFAULT_ORGANIATION=lamna
 HELM_NAMESPACE=lamna
 HELM_RELEASE=tyger
-TEST_CLIENT_CERT_FILE=~/tyger_test_client_cert.pem
+TEST_CLIENT_CERT_VERSION=9e2f7c557fbd4a2490f0cd995b2c8fca
+TEST_CLIENT_CERT_FILE=~/tyger_test_client_cert_${TEST_CLIENT_CERT_VERSION}.pem
 TEST_CLIENT_IDENTIFIER_URI=api://tyger-test-client
 AZURE_SUBSCRIPTION=BiomedicalImaging-NonProd
+TYGER_URI = https://$(shell echo '${ENVIRONMENT_CONFIG}' | jq -r '.organizations["${DEFAULT_ORGANIATION}"].subdomain').$(shell echo '${ENVIRONMENT_CONFIG}' | jq -r '.dependencies.dnsZone.name')
 
 ensure-environment:
-	scripts/get-context-environment-config.sh | deploy/scripts/ensure-environment.sh -c -
+	echo '${ENVIRONMENT_CONFIG}' | deploy/scripts/environment/ensure-environment.sh -c -
 
-remove-environment:
-	scripts/get-context-environment-config.sh | deploy/scripts/remove-environment.sh -c -
+remove-environment: down
+	echo '${ENVIRONMENT_CONFIG}' | deploy/scripts/environment/remove-environment.sh -c -
 
 set-context:
-	environment_config=$$(scripts/get-context-environment-config.sh)
-	subscription=$$(echo "$${environment_config}" | jq -r '.subscription')
-	resource_group=$$(echo "$${environment_config}" | jq -r '.resourceGroup')
-	primary_cluster_name=$$(echo "$${environment_config}" | jq -r '.primaryCluster')
+	subscription=$$(echo '${ENVIRONMENT_CONFIG}' | jq -r '.subscription')
+	resource_group=$$(echo '${ENVIRONMENT_CONFIG}' | jq -r '.resourceGroup')
+	primary_cluster_name=$$(echo '${ENVIRONMENT_CONFIG}' | jq -r '.primaryCluster')
 
 	az account set --subscription "$${subscription}"
 
 	az aks get-credentials -n "$${primary_cluster_name}" -g "$${resource_group}" --overwrite-existing
 	kubectl config set-context --current --namespace=${HELM_NAMESPACE}
-	
 
 set-localsettings:
 	if [[ $$(helm list -n "${HELM_NAMESPACE}" -l name=${HELM_RELEASE} -o json | jq length) == 0 ]]; then
 		echo "Run 'make up' before this target"; exit 1
 	fi
 
-	environment_config=$$(scripts/get-context-environment-config.sh)
-	buffer_secret_name=$$(echo "$${environment_config}" | jq -r '.organizations["${DEFAULT_ORGANIATION}"].storage.buffers[0].name')
+	buffer_secret_name=$$(echo '${ENVIRONMENT_CONFIG}' | jq -r '.organizations["${DEFAULT_ORGANIATION}"].storage.buffers[0].name')
 	buffer_secret_value="$$(kubectl get secrets -n ${HELM_NAMESPACE} $${buffer_secret_name} -o jsonpath="{.data.connectionString}" | base64 -d)"
 
-	logs_secret_name=$$(echo "$${environment_config}" | jq -r '.organizations["${DEFAULT_ORGANIATION}"].storage.logs.name')
+	logs_secret_name=$$(echo '${ENVIRONMENT_CONFIG}' | jq -r '.organizations["${DEFAULT_ORGANIATION}"].storage.logs.name')
 	logs_secret_value="$$(kubectl get secrets -n ${HELM_NAMESPACE} $${logs_secret_name} -o jsonpath="{.data.connectionString}" | base64 -d)"
 
 	postgres_password="$$(kubectl get secrets -n ${HELM_NAMESPACE} ${HELM_RELEASE}-db -o jsonpath="{.data.postgresql-password}" | base64 -d)"
@@ -48,15 +51,15 @@ set-localsettings:
 			"logging": { "Console": {"FormatterName": "simple" } },
 			"auth": {
 				"enabled": "${SECURITY_ENABLED}",
-				"authority":"$$(echo "$${environment_config}" | jq -r '.organizations["${DEFAULT_ORGANIATION}"].authority')",
-				"audience":"$$(echo "$${environment_config}" | jq -r '.organizations["${DEFAULT_ORGANIATION}"].audience')"
+				"authority":"$$(echo '${ENVIRONMENT_CONFIG}' | jq -r '.organizations["${DEFAULT_ORGANIATION}"].authority')",
+				"audience":"$$(echo '${ENVIRONMENT_CONFIG}' | jq -r '.organizations["${DEFAULT_ORGANIATION}"].audience')"
 			},
 			"kubernetes": {
 				"kubeconfigPath": "$${HOME}/.kube/config",
 				"namespace": "${HELM_NAMESPACE}",
 				"jobServiceAccount": "${HELM_RELEASE}-job",
 				"noOpConfigMap": "${HELM_RELEASE}-no-op",
-				"clusters": $$(echo "$${environment_config}" | jq -c '.clusters')
+				"clusters": $$(echo '${ENVIRONMENT_CONFIG}' | jq -c '.clusters')
 			},
 			"logArchive": {
 				"storageAccountConnectionString": "$${logs_secret_value}"
@@ -78,11 +81,11 @@ build:
 	cd ${SERVER_PATH}
 	dotnet build --no-restore
 
-run: set-localsettings check-forwarding
+run: check-forwarding set-localsettings
 	cd ${SERVER_PATH}
 	dotnet run -v m --no-restore
 
-watch: set-localsettings check-forwarding
+watch: check-forwarding set-localsettings
 	cd ${SERVER_PATH}
 	dotnet watch
 
@@ -94,23 +97,19 @@ unit-test:
 	go test ./... | { grep -v "\\[[no test files\\]" || true; }
 
 docker-build:
-	scripts/get-context-environment-config.sh | scripts/build-images.sh -c - --push --push-force --tag dev 
+	echo '${ENVIRONMENT_CONFIG}' | scripts/build-images.sh -c - --push --push-force --tag dev 
 
 docker-build-test:
-	scripts/get-context-environment-config.sh | scripts/build-images.sh -c - --test --push --push-force --tag test
+	echo '${ENVIRONMENT_CONFIG}' | scripts/build-images.sh -c - --test --push --push-force --tag test
 
-up: docker-build
-	scripts/get-context-environment-config.sh | deploy/scripts/tyger-up.sh -c -
+up: ensure-environment docker-build
+	echo '${ENVIRONMENT_CONFIG}' | deploy/scripts/tyger/tyger-up.sh -c -
 
 down:
-	scripts/get-context-environment-config.sh | deploy/scripts/tyger-down.sh -c -
+	echo '${ENVIRONMENT_CONFIG}' | deploy/scripts/tyger/tyger-down.sh -c -
 
-install-cli:
-	cd cli
-	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go install -ldflags="-s -w" -v ./cmd/tyger
-
-e2e-no-up: docker-build-test
-	if ! scripts/get-context-environment-config.sh | timeout --foreground 30m scripts/wait-for-cluster-to-scale.sh -c -; then
+e2e-no-up: docker-build-test cli-ready
+	if ! echo '${ENVIRONMENT_CONFIG}' | timeout --foreground 30m scripts/wait-for-cluster-to-scale.sh -c -; then
 		echo "timed out waiting for nodepools to scale"
 		exit 1
 	fi
@@ -118,23 +117,23 @@ e2e-no-up: docker-build-test
 	pushd cli/test/e2e
 	go test -tags=e2e
 
-	popd
-	scripts/check-login.sh
-	dvc pull
-	pytest eminence --workers 32
-
 e2e: up e2e-no-up
 
+eminence-no-up: e2e-no-up eminence-data
+	pytest eminence --workers 32
 
-test-data: 
+eminence: up eminence-no-up
+
+eminence-data: 
+	scripts/check-login.sh
+	dvc pull
 	cd scripts && ./update-gadgetron-test-data.sh
 
-test: unit-test e2e
+test: unit-test eminence
 
-test-no-up: unit-test e2e-no-up
+test-no-up: unit-test eminence-no-up
 
-
-forward:
+forward: set-context
 	scripts/forward-services.sh -n ${HELM_NAMESPACE}
 
 check-forwarding:
@@ -144,30 +143,32 @@ check-forwarding:
 	fi
 
 download-test-client-cert:
-	rm -f ${TEST_CLIENT_CERT_FILE}
-	az keyvault secret download --vault-name eminence --name tyger-test-client-cert --file ${TEST_CLIENT_CERT_FILE} --subscription ${AZURE_SUBSCRIPTION}
-	chmod 600 ${TEST_CLIENT_CERT_FILE}
+	if [[ ! -f ${TEST_CLIENT_CERT_FILE} ]]; then
+		rm -f ${TEST_CLIENT_CERT_FILE}
+		az keyvault secret download --vault-name eminence --name tyger-test-client-cert --version ${TEST_CLIENT_CERT_VERSION} --file ${TEST_CLIENT_CERT_FILE} --subscription ${AZURE_SUBSCRIPTION}
+		chmod 600 ${TEST_CLIENT_CERT_FILE}
+	fi
 
 check-test-client-cert:
 	[ -f ${TEST_CLIENT_CERT_FILE} ]
 
 get-tyger-uri:
-	environment_config=$$(scripts/get-context-environment-config.sh)
-	echo "https://$$(echo "$${environment_config}" | jq -r '.organizations["${DEFAULT_ORGANIATION}"].subdomain').$$(echo "$${environment_config}" | jq -r '.dependencies.dnsZone.name')"
+	echo ${TYGER_URI}
 
-login-service-principal:
-	environment_config=$$(scripts/get-context-environment-config.sh)
-	uri="https://$$(echo "$${environment_config}" | jq -r '.organizations["${DEFAULT_ORGANIATION}"].subdomain').$$(echo "$${environment_config}" | jq -r '.dependencies.dnsZone.name')"
+login-service-principal: install-cli download-test-client-cert
+	tyger login '${TYGER_URI}' --service-principal ${TEST_CLIENT_IDENTIFIER_URI} --cert ${TEST_CLIENT_CERT_FILE}
 
-	
-	tyger login "$${uri}" --service-principal ${TEST_CLIENT_IDENTIFIER_URI} --cert ${TEST_CLIENT_CERT_FILE}
+login: install-cli download-test-client-cert
+	tyger login "${TYGER_URI}"	
 
-login:
-	environment_config=$$(scripts/get-context-environment-config.sh)
-	uri="https://$$(echo "$${environment_config}" | jq -r '.organizations["${DEFAULT_ORGANIATION}"].subdomain').$$(echo "$${environment_config}" | jq -r '.dependencies.dnsZone.name')"
+install-cli:
+	cd cli
+	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go install -ldflags="-s -w" -v ./cmd/tyger
 
-	
-	tyger login "$${uri}"	
+cli-ready:
+	if ! tyger login status &> /dev/null; then
+		make login-service-principal
+	fi
 
 connect-db:
 	postgres_password=$(shell kubectl get secrets ${HELM_RELEASE}-db -o jsonpath="{.data.postgresql-password}" | base64 -d)
